Major project structure refactoring
