--- src/main/java/org/zalando/pazuzu/feature/FeatureRepositoryCustom.java
@@ -1,7 +1,0 @@
-package org.zalando.pazuzu.feature;
-
-import java.util.List;
-
-public interface FeatureRepositoryCustom {
-    public List<Feature> getFeatures(int offset, int limit);
-}
