+++ src/main/java/org/zalando/pazuzu/infrastructure/dto/FeatureFullDto.java
@@ -1,7 +1,7 @@
-package org.zalando.pazuzu.feature;
+package org.zalando.pazuzu.infrastructure.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.zalando.pazuzu.feature.tag.TagDto;
+import org.zalando.pazuzu.infrastructure.domain.Feature;
 
 import java.util.ArrayList;
 import java.util.List;
