+++ src/main/java/org/zalando/pazuzu/infrastructure/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package org.zalando.pazuzu.exception;
+package org.zalando.pazuzu.infrastructure.exception;
 
 public class NotFoundException extends ServiceException {
     public NotFoundException(Error error) {
