+++ src/main/java/org/zalando/pazuzu/infrastructure/exception/ServiceException.java
@@ -1,4 +1,4 @@
-package org.zalando.pazuzu.exception;
+package org.zalando.pazuzu.infrastructure.exception;
 
 
 public class ServiceException extends Exception {
