+++ src/main/java/org/zalando/pazuzu/infrastructure/resource/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package org.zalando.pazuzu.exception;
+package org.zalando.pazuzu.infrastructure.resource;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,6 +8,10 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import org.zalando.pazuzu.infrastructure.exception.BadRequestException;
+import org.zalando.pazuzu.infrastructure.exception.Error;
+import org.zalando.pazuzu.infrastructure.exception.ErrorDto;
+import org.zalando.pazuzu.infrastructure.exception.NotFoundException;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
